requirements.txt
telethon==1.29.2  # Обновлено до последней версии с поддержкой реакций
python-dotenv==1.0.0